requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdf2image>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=9.0.0
pytest-asyncio>=1.0.0
